@@ -50,7 +50,7 @@ class ArithmeticString(ArithmeticObject):
 
     @property
     def value(self):
-        return "'{}'".format(self._value)
+        return f"'{self._value}'"
 
     def __repr__(self):
         return self.value
@@ -66,7 +66,7 @@ class ArithmeticNumber(ArithmeticObject):
 
     @property
     def value(self):
-        return "{}".format(self._value)
+        return f"{self._value}"
 
     @property
     def dtype(self):
@@ -84,14 +84,16 @@ class ArithmeticSeries(ArithmeticObject):
             self._tasks = task._arithmetic_series._tasks.copy()
             self._dtype = dtype
         else:
-            aggregatable_field_name = query_compiler.display_name_to_aggregatable_name(display_name)
+            aggregatable_field_name = query_compiler.display_name_to_aggregatable_name(
+                display_name
+            )
             if aggregatable_field_name is None:
                 raise ValueError(
-                    "Can not perform arithmetic operations on non aggregatable fields"
-                    "{} is not aggregatable.".format(display_name)
+                    f"Can not perform arithmetic operations on non aggregatable fields"
+                    f"{display_name} is not aggregatable."
                 )
 
-            self._value = "doc['{}'].value".format(aggregatable_field_name)
+            self._value = f"doc['{aggregatable_field_name}'].value"
             self._tasks = []
             self._dtype = dtype
 
@@ -105,44 +107,44 @@ class ArithmeticSeries(ArithmeticObject):
 
     def __repr__(self):
         buf = StringIO()
-        buf.write("Series: {} ".format(self.value))
+        buf.write(f"Series: {self.value} ")
         buf.write("Tasks: ")
         for task in self._tasks:
-            buf.write("{} ".format(repr(task)))
+            buf.write(f"{task!r} ")
         return buf.getvalue()
 
     def resolve(self):
         value = self._value
 
         for task in self._tasks:
-            if task.op_name == '__add__':
-                value = "({} + {})".format(value, task.object.resolve())
-            elif task.op_name == '__truediv__':
-                value = "({} / {})".format(value, task.object.resolve())
-            elif task.op_name == '__floordiv__':
-                value = "Math.floor({} / {})".format(value, task.object.resolve())
-            elif task.op_name == '__mod__':
-                value = "({} % {})".format(value, task.object.resolve())
-            elif task.op_name == '__mul__':
-                value = "({} * {})".format(value, task.object.resolve())
-            elif task.op_name == '__pow__':
-                value = "Math.pow({}, {})".format(value, task.object.resolve())
-            elif task.op_name == '__sub__':
-                value = "({} - {})".format(value, task.object.resolve())
-            elif task.op_name == '__radd__':
-                value = "({} + {})".format(task.object.resolve(), value)
-            elif task.op_name == '__rtruediv__':
-                value = "({} / {})".format(task.object.resolve(), value)
-            elif task.op_name == '__rfloordiv__':
-                value = "Math.floor({} / {})".format(task.object.resolve(), value)
-            elif task.op_name == '__rmod__':
-                value = "({} % {})".format(task.object.resolve(), value)
-            elif task.op_name == '__rmul__':
-                value = "({} * {})".format(task.object.resolve(), value)
-            elif task.op_name == '__rpow__':
-                value = "Math.pow({}, {})".format(task.object.resolve(), value)
-            elif task.op_name == '__rsub__':
-                value = "({} - {})".format(task.object.resolve(), value)
+            if task.op_name == "__add__":
+                value = f"({value} + {task.object.resolve()})"
+            elif task.op_name == "__truediv__":
+                value = f"({value} / {task.object.resolve()})"
+            elif task.op_name == "__floordiv__":
+                value = f"Math.floor({value} / {task.object.resolve()})"
+            elif task.op_name == "__mod__":
+                value = f"({value} % {task.object.resolve()})"
+            elif task.op_name == "__mul__":
+                value = f"({value} * {task.object.resolve()})"
+            elif task.op_name == "__pow__":
+                value = f"Math.pow({value}, {task.object.resolve()})"
+            elif task.op_name == "__sub__":
+                value = f"({value} - {task.object.resolve()})"
+            elif task.op_name == "__radd__":
+                value = f"({task.object.resolve()} + {value})"
+            elif task.op_name == "__rtruediv__":
+                value = f"({task.object.resolve()} / {value})"
+            elif task.op_name == "__rfloordiv__":
+                value = f"Math.floor({task.object.resolve()} / {value})"
+            elif task.op_name == "__rmod__":
+                value = f"({task.object.resolve()} % {value})"
+            elif task.op_name == "__rmul__":
+                value = f"({task.object.resolve()} * {value})"
+            elif task.op_name == "__rpow__":
+                value = f"Math.pow({task.object.resolve()}, {value})"
+            elif task.op_name == "__rsub__":
+                value = f"({task.object.resolve()} - {value})"
 
         return value
 
@@ -165,20 +167,27 @@ class ArithmeticSeries(ArithmeticObject):
 
         # see end of https://pandas.pydata.org/pandas-docs/stable/getting_started/basics.html?highlight=dtype for
         # dtype heirarchy
-        if np.issubdtype(self.dtype, np.number) and np.issubdtype(right.dtype, np.number):
+        if np.issubdtype(self.dtype, np.number) and np.issubdtype(
+            right.dtype, np.number
+        ):
             # series.number op_name number (all ops)
             return True
-        elif np.issubdtype(self.dtype, np.object_) and np.issubdtype(right.dtype, np.object_):
+        elif np.issubdtype(self.dtype, np.object_) and np.issubdtype(
+            right.dtype, np.object_
+        ):
             # series.string op_name string (only add)
-            if op_name == '__add__' or op_name == '__radd__':
+            if op_name == "__add__" or op_name == "__radd__":
                 return True
-        elif np.issubdtype(self.dtype, np.object_) and np.issubdtype(right.dtype, np.integer):
+        elif np.issubdtype(self.dtype, np.object_) and np.issubdtype(
+            right.dtype, np.integer
+        ):
             # series.string op_name int (only mul)
-            if op_name == '__mul__':
+            if op_name == "__mul__":
                 return True
 
         raise TypeError(
-            "Arithmetic operation on incompatible types {} {} {}".format(self.dtype, op_name, right.dtype))
+            f"Arithmetic operation on incompatible types {self.dtype} {op_name} {right.dtype}"
+        )
 
 
 class ArithmeticTask:
@@ -186,13 +195,12 @@ class ArithmeticTask:
         self._op_name = op_name
 
         if not isinstance(object, ArithmeticObject):
-            raise TypeError("Task requires ArithmeticObject not {}".format(type(object)))
+            raise TypeError(f"Task requires ArithmeticObject not {type(object)}")
         self._object = object
 
     def __repr__(self):
         buf = StringIO()
-        buf.write("op_name: {} ".format(self.op_name))
-        buf.write("object: {} ".format(repr(self.object)))
+        buf.write(f"op_name: {self.op_name} object: {self.object!r} ")
         return buf.getvalue()
 
     @property
