@@ -24,10 +24,7 @@ import warnings
 # The license for this library can be found NOTICE.txt and the code can be
 # https://raw.githubusercontent.com/pandas-dev/pandas/v1.0.1/pandas/compat/_optional.py
 
-VERSIONS = {
-    "xgboost": "0.90",
-    "sklearn": "0.22.1"
-}
+VERSIONS = {"xgboost": "0.90", "sklearn": "0.22.1"}
 
 # Update install.rst when updating versions!
 
@@ -49,12 +46,12 @@ def _get_version(module: types.ModuleType) -> str:
         version = getattr(module, "__VERSION__", None)
 
     if version is None:
-        raise ImportError("Can't determine version for {}".format(module.__name__))
+        raise ImportError(f"Can't determine version for {module.__name__}")
     return version
 
 
 def import_optional_dependency(
-        name: str, extra: str = "", raise_on_missing: bool = True, on_version: str = "raise"
+    name: str, extra: str = "", raise_on_missing: bool = True, on_version: str = "raise"
 ):
     """
     Import an optional dependency.
