@@ -41,7 +41,16 @@ import eland.plotting
 from eland import NDFrame
 from eland.arithmetics import ArithmeticSeries, ArithmeticString, ArithmeticNumber
 from eland.common import DEFAULT_NUM_ROWS_DISPLAYED, docstring_parameter
-from eland.filter import NotFilter, Equal, Greater, Less, GreaterEqual, LessEqual, ScriptFilter, IsIn
+from eland.filter import (
+    NotFilter,
+    Equal,
+    Greater,
+    Less,
+    GreaterEqual,
+    LessEqual,
+    ScriptFilter,
+    IsIn,
+)
 
 
 def _get_method_name():
@@ -89,12 +98,14 @@ class Series(NDFrame):
     Name: Carrier, Length: 13059, dtype: object
     """
 
-    def __init__(self,
-                 client=None,
-                 index_pattern=None,
-                 name=None,
-                 index_field=None,
-                 query_compiler=None):
+    def __init__(
+        self,
+        client=None,
+        index_pattern=None,
+        name=None,
+        index_field=None,
+        query_compiler=None,
+    ):
         # Series has 1 column
         if name is None:
             columns = None
@@ -106,7 +117,8 @@ class Series(NDFrame):
             index_pattern=index_pattern,
             columns=columns,
             index_field=index_field,
-            query_compiler=query_compiler)
+            query_compiler=query_compiler,
+        )
 
     hist = eland.plotting.ed_hist_series
 
@@ -303,17 +315,17 @@ class Series(NDFrame):
 
     @docstring_parameter(DEFAULT_NUM_ROWS_DISPLAYED)
     def to_string(
-            self,
-            buf=None,
-            na_rep="NaN",
-            float_format=None,
-            header=True,
-            index=True,
-            length=False,
-            dtype=False,
-            name=False,
-            max_rows=None,
-            min_rows=None,
+        self,
+        buf=None,
+        na_rep="NaN",
+        float_format=None,
+        header=True,
+        index=True,
+        length=False,
+        dtype=False,
+        name=False,
+        max_rows=None,
+        min_rows=None,
     ):
         """
         Render a string representation of the Series.
@@ -335,12 +347,13 @@ class Series(NDFrame):
             else:
                 max_rows = min(num_rows, max_rows)
         elif max_rows is None:
-            warnings.warn("Series.to_string called without max_rows set "
-                          "- this will return entire index results. "
-                          "Setting max_rows={default}"
-                          " overwrite if different behaviour is required."
-                          .format(default=DEFAULT_NUM_ROWS_DISPLAYED),
-                          UserWarning)
+            warnings.warn(
+                f"Series.to_string called without max_rows set "
+                f"- this will return entire index results. "
+                f"Setting max_rows={DEFAULT_NUM_ROWS_DISPLAYED}"
+                f" overwrite if different behaviour is required.",
+                UserWarning,
+            )
             max_rows = DEFAULT_NUM_ROWS_DISPLAYED
 
         # because of the way pandas handles max_rows=0, not having this throws an error
@@ -357,31 +370,29 @@ class Series(NDFrame):
             _buf = StringIO()
 
         # Create repr of fake series without name, length, dtype summary
-        temp_str = temp_series.to_string(buf=_buf,
-                                         na_rep=na_rep,
-                                         float_format=float_format,
-                                         header=header,
-                                         index=index,
-                                         length=False,
-                                         dtype=False,
-                                         name=False,
-                                         max_rows=max_rows)
+        temp_series.to_string(
+            buf=_buf,
+            na_rep=na_rep,
+            float_format=float_format,
+            header=header,
+            index=index,
+            length=False,
+            dtype=False,
+            name=False,
+            max_rows=max_rows,
+        )
 
         # Create the summary
-        footer = ""
+        footer = []
         if name and self.name is not None:
-            footer += "Name: {}".format(str(self.name))
+            footer.append(f"Name: {self.name}")
         if length and len(self) > max_rows:
-            if footer:
-                footer += ", "
-            footer += "Length: {}".format(len(self.index))
+            footer.append(f"Length: {len(self.index)}")
         if dtype:
-            if footer:
-                footer += ", "
-            footer += "dtype: {}".format(temp_series.dtype)
+            footer.append(f"dtype: {temp_series.dtype}")
 
-        if len(footer) > 0:
-            _buf.write("\n{}".format(footer))
+        if footer:
+            _buf.write(f"\n{', '.join(footer)}")
 
         if buf is None:
             result = _buf.getvalue()
@@ -398,7 +409,7 @@ class Series(NDFrame):
     def __gt__(self, other):
         if isinstance(other, Series):
             # Need to use scripted query to compare to values
-            painless = "doc['{}'].value > doc['{}'].value".format(self.name, other.name)
+            painless = f"doc['{self.name}'].value > doc['{other.name}'].value"
             return ScriptFilter(painless)
         elif isinstance(other, (int, float)):
             return Greater(field=self.name, value=other)
@@ -408,7 +419,7 @@ class Series(NDFrame):
     def __lt__(self, other):
         if isinstance(other, Series):
             # Need to use scripted query to compare to values
-            painless = "doc['{}'].value < doc['{}'].value".format(self.name, other.name)
+            painless = f"doc['{self.name}'].value < doc['{other.name}'].value"
             return ScriptFilter(painless)
         elif isinstance(other, (int, float)):
             return Less(field=self.name, value=other)
@@ -418,7 +429,7 @@ class Series(NDFrame):
     def __ge__(self, other):
         if isinstance(other, Series):
             # Need to use scripted query to compare to values
-            painless = "doc['{}'].value >= doc['{}'].value".format(self.name, other.name)
+            painless = f"doc['{self.name}'].value >= doc['{other.name}'].value"
             return ScriptFilter(painless)
         elif isinstance(other, (int, float)):
             return GreaterEqual(field=self.name, value=other)
@@ -428,7 +439,7 @@ class Series(NDFrame):
     def __le__(self, other):
         if isinstance(other, Series):
             # Need to use scripted query to compare to values
-            painless = "doc['{}'].value <= doc['{}'].value".format(self.name, other.name)
+            painless = f"doc['{self.name}'].value <= doc['{other.name}'].value"
             return ScriptFilter(painless)
         elif isinstance(other, (int, float)):
             return LessEqual(field=self.name, value=other)
@@ -438,7 +449,7 @@ class Series(NDFrame):
     def __eq__(self, other):
         if isinstance(other, Series):
             # Need to use scripted query to compare to values
-            painless = "doc['{}'].value == doc['{}'].value".format(self.name, other.name)
+            painless = f"doc['{self.name}'].value == doc['{other.name}'].value"
             return ScriptFilter(painless)
         elif isinstance(other, (int, float)):
             return Equal(field=self.name, value=other)
@@ -450,7 +461,7 @@ class Series(NDFrame):
     def __ne__(self, other):
         if isinstance(other, Series):
             # Need to use scripted query to compare to values
-            painless = "doc['{}'].value != doc['{}'].value".format(self.name, other.name)
+            painless = f"doc['{self.name}'].value != doc['{other.name}'].value"
             return ScriptFilter(painless)
         elif isinstance(other, (int, float)):
             return NotFilter(Equal(field=self.name, value=other))
@@ -1057,7 +1068,9 @@ class Series(NDFrame):
             # Check we can the 2 Series are compatible (raises on error):
             self._query_compiler.check_arithmetics(right._query_compiler)
 
-            right_object = ArithmeticSeries(right._query_compiler, right.name, right._dtype)
+            right_object = ArithmeticSeries(
+                right._query_compiler, right.name, right._dtype
+            )
             display_name = None
         elif np.issubdtype(np.dtype(type(right)), np.number):
             right_object = ArithmeticNumber(right, np.dtype(type(right)))
@@ -1067,14 +1080,19 @@ class Series(NDFrame):
             display_name = self.name
         else:
             raise TypeError(
-                "unsupported operation type(s) ['{}'] for operands ['{}' with dtype '{}', '{}']"
-                    .format(method_name, type(self), self._dtype, type(right).__name__)
+                f"unsupported operation type(s) [{method_name!r}] "
+                f"for operands ['{type(self)}' with dtype '{self._dtype}', "
+                f"'{type(right).__name__}']"
             )
 
         left_object = ArithmeticSeries(self._query_compiler, self.name, self._dtype)
         left_object.arithmetic_operation(method_name, right_object)
 
-        series = Series(query_compiler=self._query_compiler.arithmetic_op_fields(display_name, left_object))
+        series = Series(
+            query_compiler=self._query_compiler.arithmetic_op_fields(
+                display_name, left_object
+            )
+        )
 
         # force set name to 'display_name'
         series._query_compiler._mappings.display_names = [display_name]
@@ -1220,7 +1238,7 @@ class Series(NDFrame):
         --------
         >>> ed_s = ed.Series('localhost', 'flights', name='Carrier').head(5)
         >>> pd_s = ed.eland_to_pandas(ed_s)
-        >>> print("type(ed_s)={0}\\ntype(pd_s)={1}".format(type(ed_s), type(pd_s)))
+        >>> print(f"type(ed_s)={type(ed_s)}\\ntype(pd_s)={type(pd_s)}")
         type(ed_s)=<class 'eland.series.Series'>
         type(pd_s)=<class 'pandas.core.series.Series'>
         >>> ed_s
