@@ -12,9 +12,11 @@
 #      See the License for the specific language governing permissions and
 #      limitations under the License.
 
-__title__ = 'eland'
-__description__ = 'Python elasticsearch client to analyse, explore and manipulate data that resides in elasticsearch.'
-__url__ = 'https://github.com/elastic/eland'
-__version__ = '7.6.0a4'
-__maintainer__ = 'Elasticsearch B.V.'
-__maintainer_email__ = 'steve.dodson@elastic.co'
+__title__ = "eland"
+__description__ = "Python elasticsearch client to analyse, explore and manipulate data that resides in elasticsearch."
+__url__ = "https://github.com/elastic/eland"
+__version__ = "7.6.0a4"
+__author__ = "Steve Dodson"
+__author_email__ = "steve.dodson@elastic.co"
+__maintainer__ = "Seth Michael Larson"
+__maintainer_email__ = "seth.larson@elastic.co"
