@@ -15,13 +15,10 @@
 # File called _pytest for PyCharm compatability
 
 from eland.tests.common import TestData
-from eland.tests.common import (
-    assert_pandas_eland_frame_equal
-)
+from eland.tests.common import assert_pandas_eland_frame_equal
 
 
 class TestDataFrameDrop(TestData):
-
     def test_flights_small_drop(self):
         ed_flights_small = self.ed_flights_small()
         pd_flights_small = self.pd_flights_small()
@@ -33,17 +30,17 @@ class TestDataFrameDrop(TestData):
         #        'FlightTimeMin', 'Origin', 'OriginAirportID', 'OriginCityName',
         #        'OriginCountry', 'OriginLocation', 'OriginRegion', 'OriginWeather',
         #        'dayOfWeek', 'timestamp']
-        pd_col0 = pd_flights_small.drop(['Carrier', 'DestCityName'], axis=1)
-        pd_col1 = pd_flights_small.drop(columns=['Carrier', 'DestCityName'])
+        pd_col0 = pd_flights_small.drop(["Carrier", "DestCityName"], axis=1)
+        pd_col1 = pd_flights_small.drop(columns=["Carrier", "DestCityName"])
 
-        ed_col0 = ed_flights_small.drop(['Carrier', 'DestCityName'], axis=1)
-        ed_col1 = ed_flights_small.drop(columns=['Carrier', 'DestCityName'])
+        ed_col0 = ed_flights_small.drop(["Carrier", "DestCityName"], axis=1)
+        ed_col1 = ed_flights_small.drop(columns=["Carrier", "DestCityName"])
 
         assert_pandas_eland_frame_equal(pd_col0, ed_col0)
         assert_pandas_eland_frame_equal(pd_col1, ed_col1)
 
         # Drop rows by index
-        pd_idx0 = pd_flights_small.drop(['1', '2'])
-        ed_idx0 = ed_flights_small.drop(['1', '2'])
+        pd_idx0 = pd_flights_small.drop(["1", "2"])
+        ed_idx0 = ed_flights_small.drop(["1", "2"])
 
         assert_pandas_eland_frame_equal(pd_idx0, ed_idx0)
