@@ -17,27 +17,28 @@
 from eland.tests.common import TestData
 from eland.tests.common import (
     assert_pandas_eland_frame_equal,
-    assert_pandas_eland_series_equal
+    assert_pandas_eland_series_equal,
 )
 
 
 class TestDataFrameGetItem(TestData):
-
     def test_getitem_one_attribute(self):
         ed_flights = self.ed_flights().head(103)
         pd_flights = self.pd_flights().head(103)
 
-        ed_flights_OriginAirportID = ed_flights['OriginAirportID']
-        pd_flights_OriginAirportID = pd_flights['OriginAirportID']
+        ed_flights_OriginAirportID = ed_flights["OriginAirportID"]
+        pd_flights_OriginAirportID = pd_flights["OriginAirportID"]
 
-        assert_pandas_eland_series_equal(pd_flights_OriginAirportID, ed_flights_OriginAirportID)
+        assert_pandas_eland_series_equal(
+            pd_flights_OriginAirportID, ed_flights_OriginAirportID
+        )
 
     def test_getitem_attribute_list(self):
         ed_flights = self.ed_flights().head(42)
         pd_flights = self.pd_flights().head(42)
 
-        ed_flights_slice = ed_flights[['OriginAirportID', 'AvgTicketPrice', 'Carrier']]
-        pd_flights_slice = pd_flights[['OriginAirportID', 'AvgTicketPrice', 'Carrier']]
+        ed_flights_slice = ed_flights[["OriginAirportID", "AvgTicketPrice", "Carrier"]]
+        pd_flights_slice = pd_flights[["OriginAirportID", "AvgTicketPrice", "Carrier"]]
 
         assert_pandas_eland_frame_equal(pd_flights_slice, ed_flights_slice)
 
@@ -48,19 +49,23 @@ class TestDataFrameGetItem(TestData):
         ed_flights_OriginAirportID = ed_flights.OriginAirportID
         pd_flights_OriginAirportID = pd_flights.OriginAirportID
 
-        assert_pandas_eland_series_equal(pd_flights_OriginAirportID, ed_flights_OriginAirportID)
+        assert_pandas_eland_series_equal(
+            pd_flights_OriginAirportID, ed_flights_OriginAirportID
+        )
 
     def test_getitem_multiple_calls(self):
         ed_flights = self.ed_flights().head(89)
         pd_flights = self.pd_flights().head(89)
 
-        ed_col0 = ed_flights[['DestCityName', 'DestCountry', 'DestLocation', 'DestRegion']]
+        ed_col0 = ed_flights[
+            ["DestCityName", "DestCountry", "DestLocation", "DestRegion"]
+        ]
         try:
-            ed_col1 = ed_col0['Carrier']
+            ed_col1 = ed_col0["Carrier"]
         except KeyError:
             pass
 
-        pd_col1 = pd_flights['DestCountry']
-        ed_col1 = ed_col0['DestCountry']
+        pd_col1 = pd_flights["DestCountry"]
+        ed_col1 = ed_col0["DestCountry"]
 
         assert_pandas_eland_series_equal(pd_col1, ed_col1)
