@@ -22,7 +22,6 @@ from eland.tests.common import TestData
 
 
 class TestDataFrameHist(TestData):
-
     def test_flights_hist(self):
         pd_flights = self.pd_flights()
         ed_flights = self.ed_flights()
@@ -30,17 +29,27 @@ class TestDataFrameHist(TestData):
         num_bins = 10
 
         # pandas data
-        pd_distancekilometers = np.histogram(pd_flights['DistanceKilometers'], num_bins)
-        pd_flightdelaymin = np.histogram(pd_flights['FlightDelayMin'], num_bins)
+        pd_distancekilometers = np.histogram(pd_flights["DistanceKilometers"], num_bins)
+        pd_flightdelaymin = np.histogram(pd_flights["FlightDelayMin"], num_bins)
 
         pd_bins = pd.DataFrame(
-            {'DistanceKilometers': pd_distancekilometers[1], 'FlightDelayMin': pd_flightdelaymin[1]})
+            {
+                "DistanceKilometers": pd_distancekilometers[1],
+                "FlightDelayMin": pd_flightdelaymin[1],
+            }
+        )
         pd_weights = pd.DataFrame(
-            {'DistanceKilometers': pd_distancekilometers[0], 'FlightDelayMin': pd_flightdelaymin[0]})
+            {
+                "DistanceKilometers": pd_distancekilometers[0],
+                "FlightDelayMin": pd_flightdelaymin[0],
+            }
+        )
 
-        t = ed_flights[['DistanceKilometers', 'FlightDelayMin']]
+        _ = ed_flights[["DistanceKilometers", "FlightDelayMin"]]
 
-        ed_bins, ed_weights = ed_flights[['DistanceKilometers', 'FlightDelayMin']]._hist(num_bins=num_bins)
+        ed_bins, ed_weights = ed_flights[
+            ["DistanceKilometers", "FlightDelayMin"]
+        ]._hist(num_bins=num_bins)
 
         # Numbers are slightly different
         assert_almost_equal(pd_bins, ed_bins)
@@ -56,17 +65,27 @@ class TestDataFrameHist(TestData):
         num_bins = 10
 
         # pandas data
-        pd_distancekilometers = np.histogram(pd_flights['DistanceKilometers'], num_bins)
-        pd_flightdelaymin = np.histogram(pd_flights['FlightDelayMin'], num_bins)
+        pd_distancekilometers = np.histogram(pd_flights["DistanceKilometers"], num_bins)
+        pd_flightdelaymin = np.histogram(pd_flights["FlightDelayMin"], num_bins)
 
         pd_bins = pd.DataFrame(
-            {'DistanceKilometers': pd_distancekilometers[1], 'FlightDelayMin': pd_flightdelaymin[1]})
+            {
+                "DistanceKilometers": pd_distancekilometers[1],
+                "FlightDelayMin": pd_flightdelaymin[1],
+            }
+        )
         pd_weights = pd.DataFrame(
-            {'DistanceKilometers': pd_distancekilometers[0], 'FlightDelayMin': pd_flightdelaymin[0]})
+            {
+                "DistanceKilometers": pd_distancekilometers[0],
+                "FlightDelayMin": pd_flightdelaymin[0],
+            }
+        )
 
-        t = ed_flights[['DistanceKilometers', 'FlightDelayMin']]
+        _ = ed_flights[["DistanceKilometers", "FlightDelayMin"]]
 
-        ed_bins, ed_weights = ed_flights[['DistanceKilometers', 'FlightDelayMin']]._hist(num_bins=num_bins)
+        ed_bins, ed_weights = ed_flights[
+            ["DistanceKilometers", "FlightDelayMin"]
+        ]._hist(num_bins=num_bins)
 
         # Numbers are slightly different
         assert_almost_equal(pd_bins, ed_bins)
