@@ -23,12 +23,13 @@ from eland.tests.common import assert_pandas_eland_frame_equal
 
 
 class TestDataFrameQuery(TestData):
-
     def test_getitem_query(self):
         # Examples from:
         # https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.query.html
-        pd_df = pd.DataFrame({'A': range(1, 6), 'B': range(10, 0, -2), 'C': range(10, 5, -1)},
-                             index=['0', '1', '2', '3', '4'])
+        pd_df = pd.DataFrame(
+            {"A": range(1, 6), "B": range(10, 0, -2), "C": range(10, 5, -1)},
+            index=["0", "1", "2", "3", "4"],
+        )
         """
         >>> pd_df
            A   B   C
@@ -39,9 +40,11 @@ class TestDataFrameQuery(TestData):
         4  5   2   6
         """
         # Now create index
-        index_name = 'eland_test_query'
+        index_name = "eland_test_query"
 
-        ed_df = ed.pandas_to_eland(pd_df, ES_TEST_CLIENT, index_name, es_if_exists="replace", es_refresh=True)
+        ed_df = ed.pandas_to_eland(
+            pd_df, ES_TEST_CLIENT, index_name, es_if_exists="replace", es_refresh=True
+        )
 
         assert_pandas_eland_frame_equal(pd_df, ed_df)
 
@@ -71,20 +74,27 @@ class TestDataFrameQuery(TestData):
         ed_flights = self.ed_flights()
         pd_flights = self.pd_flights()
 
-        assert pd_flights.query('FlightDelayMin > 60').shape == ed_flights.query('FlightDelayMin > 60').shape
+        assert (
+            pd_flights.query("FlightDelayMin > 60").shape
+            == ed_flights.query("FlightDelayMin > 60").shape
+        )
 
     def test_isin_query(self):
         ed_flights = self.ed_flights()
         pd_flights = self.pd_flights()
 
-        assert pd_flights[pd_flights.OriginAirportID.isin(['LHR', 'SYD'])].shape == \
-               ed_flights[ed_flights.OriginAirportID.isin(['LHR', 'SYD'])].shape
+        assert (
+            pd_flights[pd_flights.OriginAirportID.isin(["LHR", "SYD"])].shape
+            == ed_flights[ed_flights.OriginAirportID.isin(["LHR", "SYD"])].shape
+        )
 
     def test_multiitem_query(self):
         # Examples from:
         # https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.query.html
-        pd_df = pd.DataFrame({'A': range(1, 6), 'B': range(10, 0, -2), 'C': range(10, 5, -1)},
-                             index=['0', '1', '2', '3', '4'])
+        pd_df = pd.DataFrame(
+            {"A": range(1, 6), "B": range(10, 0, -2), "C": range(10, 5, -1)},
+            index=["0", "1", "2", "3", "4"],
+        )
         """
         >>> pd_df
            A   B   C
@@ -95,9 +105,11 @@ class TestDataFrameQuery(TestData):
         4  5   2   6
         """
         # Now create index
-        index_name = 'eland_test_query'
+        index_name = "eland_test_query"
 
-        ed_df = ed.pandas_to_eland(pd_df, ES_TEST_CLIENT, index_name, es_if_exists="replace", es_refresh=True)
+        ed_df = ed.pandas_to_eland(
+            pd_df, ES_TEST_CLIENT, index_name, es_if_exists="replace", es_refresh=True
+        )
 
         assert_pandas_eland_frame_equal(pd_df, ed_df)
 
@@ -116,14 +128,14 @@ class TestDataFrameQuery(TestData):
         assert_pandas_eland_frame_equal(pd_q2, ed_q2)
         assert_pandas_eland_frame_equal(pd_q3, ed_q3)
 
-        ed_q4 = ed_q1.query('B > 2')
-        pd_q4 = pd_q1.query('B > 2')
+        ed_q4 = ed_q1.query("B > 2")
+        pd_q4 = pd_q1.query("B > 2")
 
         assert_pandas_eland_frame_equal(pd_q4, ed_q4)
 
         # Drop rows by index
-        ed_q4 = ed_q4.drop(['2'])
-        pd_q4 = pd_q4.drop(['2'])
+        ed_q4 = ed_q4.drop(["2"])
+        pd_q4 = pd_q4.drop(["2"])
 
         assert_pandas_eland_frame_equal(pd_q4, ed_q4)
 
