@@ -17,30 +17,15 @@
 import pandas as pd
 import pytest
 
-from eland.compat import PY36
 from eland.dataframe import DEFAULT_NUM_ROWS_DISPLAYED
 from eland.tests.common import TestData, assert_pandas_eland_series_equal
 
 
 class TestDataFrameRepr(TestData):
-
     @classmethod
     def setup_class(cls):
         # conftest.py changes this default - restore to original setting
-        pd.set_option('display.max_rows', 60)
-
-    # Override these methods as:
-    # {'lat': '-33.94609833', 'lon': '151.177002'} order is not consistent in python 3.5 (dict's not ordered)
-    # remove from test for now
-    def ed_flights(self):
-        if not PY36:
-            return super().ed_flights().drop(columns=['OriginLocation', 'DestLocation'])
-        return super().ed_flights()
-
-    def pd_flights(self):
-        if not PY36:
-            return super().pd_flights().drop(columns=['OriginLocation', 'DestLocation'])
-        return super().pd_flights()
+        pd.set_option("display.max_rows", 60)
 
     """
     to_string
@@ -81,40 +66,34 @@ class TestDataFrameRepr(TestData):
 
         Hence we store the pandas df source json as 'lon', 'lat'
         """
-        if PY36:
-            pd_dest_location = self.pd_flights()['DestLocation'].head(1)
-            ed_dest_location = self.ed_flights()['DestLocation'].head(1)
+        pd_dest_location = self.pd_flights()["DestLocation"].head(1)
+        ed_dest_location = self.ed_flights()["DestLocation"].head(1)
 
-            assert_pandas_eland_series_equal(pd_dest_location, ed_dest_location)
-        else:
-            # NOOP
-            assert True
+        assert_pandas_eland_series_equal(pd_dest_location, ed_dest_location)
 
     def test_num_rows_to_string(self):
-        if PY36:
-            # check setup works
-            assert pd.get_option('display.max_rows') == 60
+        # check setup works
+        assert pd.get_option("display.max_rows") == 60
 
-            # Test eland.DataFrame.to_string vs pandas.DataFrame.to_string
-            # In pandas calling 'to_string' without max_rows set, will dump ALL rows
+        # Test eland.DataFrame.to_string vs pandas.DataFrame.to_string
+        # In pandas calling 'to_string' without max_rows set, will dump ALL rows
 
-            # Test n-1, n, n+1 for edge cases
-            self.num_rows_to_string(DEFAULT_NUM_ROWS_DISPLAYED - 1)
-            self.num_rows_to_string(DEFAULT_NUM_ROWS_DISPLAYED)
-            with pytest.warns(UserWarning):
-                # UserWarning displayed by eland here (compare to pandas with max_rows set)
-                self.num_rows_to_string(DEFAULT_NUM_ROWS_DISPLAYED + 1, None, DEFAULT_NUM_ROWS_DISPLAYED)
+        # Test n-1, n, n+1 for edge cases
+        self.num_rows_to_string(DEFAULT_NUM_ROWS_DISPLAYED - 1)
+        self.num_rows_to_string(DEFAULT_NUM_ROWS_DISPLAYED)
+        with pytest.warns(UserWarning):
+            # UserWarning displayed by eland here (compare to pandas with max_rows set)
+            self.num_rows_to_string(
+                DEFAULT_NUM_ROWS_DISPLAYED + 1, None, DEFAULT_NUM_ROWS_DISPLAYED
+            )
 
-            # Test for where max_rows lt or gt num_rows
-            self.num_rows_to_string(10, 5, 5)
-            self.num_rows_to_string(100, 200, 200)
-        else:
-            # NOOP
-            assert True
+        # Test for where max_rows lt or gt num_rows
+        self.num_rows_to_string(10, 5, 5)
+        self.num_rows_to_string(100, 200, 200)
 
     def num_rows_to_string(self, rows, max_rows_eland=None, max_rows_pandas=None):
-        ed_flights = self.ed_flights()[['DestLocation', 'OriginLocation']]
-        pd_flights = self.pd_flights()[['DestLocation', 'OriginLocation']]
+        ed_flights = self.ed_flights()[["DestLocation", "OriginLocation"]]
+        pd_flights = self.pd_flights()[["DestLocation", "OriginLocation"]]
 
         ed_head = ed_flights.head(rows)
         pd_head = pd_flights.head(rows)
@@ -131,8 +110,8 @@ class TestDataFrameRepr(TestData):
         ed_ecom = self.ed_ecommerce()
         pd_ecom = self.pd_ecommerce()
 
-        ed_ecom_s = ed_ecom[ed_ecom['currency'] == 'USD'].to_string()
-        pd_ecom_s = pd_ecom[pd_ecom['currency'] == 'USD'].to_string()
+        ed_ecom_s = ed_ecom[ed_ecom["currency"] == "USD"].to_string()
+        pd_ecom_s = pd_ecom[pd_ecom["currency"] == "USD"].to_string()
 
         assert ed_ecom_s == pd_ecom_s
 
@@ -141,9 +120,15 @@ class TestDataFrameRepr(TestData):
     """
 
     def test_num_rows_repr(self):
-        self.num_rows_repr(pd.get_option('display.max_rows') - 1, pd.get_option('display.max_rows') - 1)
-        self.num_rows_repr(pd.get_option('display.max_rows'), pd.get_option('display.max_rows'))
-        self.num_rows_repr(pd.get_option('display.max_rows') + 1, pd.get_option('display.min_rows'))
+        self.num_rows_repr(
+            pd.get_option("display.max_rows") - 1, pd.get_option("display.max_rows") - 1
+        )
+        self.num_rows_repr(
+            pd.get_option("display.max_rows"), pd.get_option("display.max_rows")
+        )
+        self.num_rows_repr(
+            pd.get_option("display.max_rows") + 1, pd.get_option("display.min_rows")
+        )
 
     def num_rows_repr(self, rows, num_rows_printed):
         ed_flights = self.ed_flights()
@@ -168,8 +153,8 @@ class TestDataFrameRepr(TestData):
         ed_ecom = self.ed_ecommerce()
         pd_ecom = self.pd_ecommerce()
 
-        ed_ecom_r = repr(ed_ecom[ed_ecom['currency'] == 'USD'])
-        pd_ecom_r = repr(pd_ecom[pd_ecom['currency'] == 'USD'])
+        ed_ecom_r = repr(ed_ecom[ed_ecom["currency"] == "USD"])
+        pd_ecom_r = repr(pd_ecom[pd_ecom["currency"] == "USD"])
 
         assert ed_ecom_r == pd_ecom_r
 
@@ -179,7 +164,7 @@ class TestDataFrameRepr(TestData):
 
     def test_num_rows_to_html(self):
         # check setup works
-        assert pd.get_option('display.max_rows') == 60
+        assert pd.get_option("display.max_rows") == 60
 
         # Test eland.DataFrame.to_string vs pandas.DataFrame.to_string
         # In pandas calling 'to_string' without max_rows set, will dump ALL rows
@@ -189,7 +174,9 @@ class TestDataFrameRepr(TestData):
         self.num_rows_to_html(DEFAULT_NUM_ROWS_DISPLAYED)
         with pytest.warns(UserWarning):
             # UserWarning displayed by eland here
-            self.num_rows_to_html(DEFAULT_NUM_ROWS_DISPLAYED + 1, None, DEFAULT_NUM_ROWS_DISPLAYED)
+            self.num_rows_to_html(
+                DEFAULT_NUM_ROWS_DISPLAYED + 1, None, DEFAULT_NUM_ROWS_DISPLAYED
+            )
 
         # Test for where max_rows lt or gt num_rows
         self.num_rows_to_html(10, 5, 5)
@@ -214,8 +201,8 @@ class TestDataFrameRepr(TestData):
         ed_ecom = self.ed_ecommerce()
         pd_ecom = self.pd_ecommerce()
 
-        ed_ecom_h = ed_ecom[ed_ecom['currency'] == 'USD'].to_html()
-        pd_ecom_h = pd_ecom[pd_ecom['currency'] == 'USD'].to_html()
+        ed_ecom_h = ed_ecom[ed_ecom["currency"] == "USD"].to_html()
+        pd_ecom_h = pd_ecom[pd_ecom["currency"] == "USD"].to_html()
 
         assert ed_ecom_h == pd_ecom_h
 
@@ -225,24 +212,26 @@ class TestDataFrameRepr(TestData):
 
     def test_num_rows_repr_html(self):
         # check setup works
-        assert pd.get_option('display.max_rows') == 60
+        assert pd.get_option("display.max_rows") == 60
 
-        show_dimensions = pd.get_option('display.show_dimensions')
+        show_dimensions = pd.get_option("display.show_dimensions")
 
         # TODO - there is a bug in 'show_dimensions' as it gets added after the last </div>
         # For now test without this
-        pd.set_option('display.show_dimensions', False)
+        pd.set_option("display.show_dimensions", False)
 
         # Test eland.DataFrame.to_string vs pandas.DataFrame.to_string
         # In pandas calling 'to_string' without max_rows set, will dump ALL rows
 
         # Test n-1, n, n+1 for edge cases
-        self.num_rows_repr_html(pd.get_option('display.max_rows') - 1)
-        self.num_rows_repr_html(pd.get_option('display.max_rows'))
-        self.num_rows_repr_html(pd.get_option('display.max_rows') + 1, pd.get_option('display.max_rows'))
+        self.num_rows_repr_html(pd.get_option("display.max_rows") - 1)
+        self.num_rows_repr_html(pd.get_option("display.max_rows"))
+        self.num_rows_repr_html(
+            pd.get_option("display.max_rows") + 1, pd.get_option("display.max_rows")
+        )
 
         # Restore default
-        pd.set_option('display.show_dimensions', show_dimensions)
+        pd.set_option("display.show_dimensions", show_dimensions)
 
     def num_rows_repr_html(self, rows, max_rows=None):
         ed_flights = self.ed_flights()
@@ -257,24 +246,21 @@ class TestDataFrameRepr(TestData):
         # print(ed_head_str)
         # print(pd_head_str)
 
-        # Currently pandas display bold_rows=True with >=PY36 and bold_rows=False with 3.5
-        # TODO - fix this test for 3.5
-        if PY36:
-            assert pd_head_str == ed_head_str
+        assert pd_head_str == ed_head_str
 
     def test_empty_dataframe_repr_html(self):
         # TODO - there is a bug in 'show_dimensions' as it gets added after the last </div>
         # For now test without this
-        show_dimensions = pd.get_option('display.show_dimensions')
-        pd.set_option('display.show_dimensions', False)
+        show_dimensions = pd.get_option("display.show_dimensions")
+        pd.set_option("display.show_dimensions", False)
 
         ed_ecom = self.ed_ecommerce()
         pd_ecom = self.pd_ecommerce()
 
-        ed_ecom_rh = ed_ecom[ed_ecom['currency'] == 'USD']._repr_html_()
-        pd_ecom_rh = pd_ecom[pd_ecom['currency'] == 'USD']._repr_html_()
+        ed_ecom_rh = ed_ecom[ed_ecom["currency"] == "USD"]._repr_html_()
+        pd_ecom_rh = pd_ecom[pd_ecom["currency"] == "USD"]._repr_html_()
 
         # Restore default
-        pd.set_option('display.show_dimensions', show_dimensions)
+        pd.set_option("display.show_dimensions", show_dimensions)
 
         assert ed_ecom_rh == pd_ecom_rh
