@@ -28,24 +28,28 @@ from eland.tests.common import TestData
 
 
 class TestDataFrameToCSV(TestData):
-
     def test_to_csv_head(self):
-        results_file = ROOT_DIR + '/dataframe/results/test_to_csv_head.csv'
+        results_file = ROOT_DIR + "/dataframe/results/test_to_csv_head.csv"
 
         ed_flights = self.ed_flights().head()
         pd_flights = self.pd_flights().head()
         ed_flights.to_csv(results_file)
         # Converting back from csv is messy as pd_flights is created from a json file
-        pd_from_csv = pd.read_csv(results_file, index_col=0, converters={
-            'DestLocation': lambda x: ast.literal_eval(x),
-            'OriginLocation': lambda x: ast.literal_eval(x)})
+        pd_from_csv = pd.read_csv(
+            results_file,
+            index_col=0,
+            converters={
+                "DestLocation": lambda x: ast.literal_eval(x),
+                "OriginLocation": lambda x: ast.literal_eval(x),
+            },
+        )
         pd_from_csv.index = pd_from_csv.index.map(str)
         pd_from_csv.timestamp = pd.to_datetime(pd_from_csv.timestamp)
 
         assert_frame_equal(pd_flights, pd_from_csv)
 
     def test_to_csv_full(self):
-        results_file = ROOT_DIR + '/dataframe/results/test_to_csv_full.csv'
+        results_file = ROOT_DIR + "/dataframe/results/test_to_csv_full.csv"
 
         # Test is slow as it's for the full dataset, but it is useful as it goes over 10000 docs
         ed_flights = self.ed_flights()
@@ -53,9 +57,14 @@ class TestDataFrameToCSV(TestData):
 
         ed_flights.to_csv(results_file)
         # Converting back from csv is messy as pd_flights is created from a json file
-        pd_from_csv = pd.read_csv(results_file, index_col=0, converters={
-            'DestLocation': lambda x: ast.literal_eval(x),
-            'OriginLocation': lambda x: ast.literal_eval(x)})
+        pd_from_csv = pd.read_csv(
+            results_file,
+            index_col=0,
+            converters={
+                "DestLocation": lambda x: ast.literal_eval(x),
+                "OriginLocation": lambda x: ast.literal_eval(x),
+            },
+        )
         pd_from_csv.index = pd_from_csv.index.map(str)
         pd_from_csv.timestamp = pd.to_datetime(pd_from_csv.timestamp)
 
@@ -64,14 +73,20 @@ class TestDataFrameToCSV(TestData):
         # Now read the csv to an index
         now_millis = int(round(time.time() * 1000))
 
-        test_index = FLIGHTS_INDEX_NAME + '.' + str(now_millis)
+        test_index = FLIGHTS_INDEX_NAME + "." + str(now_millis)
 
-        ed_flights_from_csv = ed.read_csv(results_file, ES_TEST_CLIENT, test_index, index_col=0, es_refresh=True,
-                                          es_geo_points=['OriginLocation', 'DestLocation'],
-                                          converters={
-                                              'DestLocation': lambda x: ast.literal_eval(x),
-                                              'OriginLocation': lambda x: ast.literal_eval(x)}
-                                          )
+        ed_flights_from_csv = ed.read_csv(
+            results_file,
+            ES_TEST_CLIENT,
+            test_index,
+            index_col=0,
+            es_refresh=True,
+            es_geo_points=["OriginLocation", "DestLocation"],
+            converters={
+                "DestLocation": lambda x: ast.literal_eval(x),
+                "OriginLocation": lambda x: ast.literal_eval(x),
+            },
+        )
         pd_flights_from_csv = ed.eland_to_pandas(ed_flights_from_csv)
 
         # TODO - there is a 'bug' where the Elasticsearch index returns data in a different order to the CSV
