@@ -12,8 +12,6 @@
 #      See the License for the specific language governing permissions and
 #      limitations under the License.
 
-from collections import OrderedDict
-
 # File called _pytest for PyCharm compatability
 import pytest
 
@@ -22,87 +20,88 @@ from eland.tests.common import TestData
 
 
 class TestMappingsWithType(TestData):
-
     def test_mappings_with_type(self):
         # Unless we spin up a 6.x index, this is difficult
         # to test. This is not ideal, but supporting some basic
         # features on 6.x indices makes eland more generally usable.
         #
         # For now, just test function:
-        mapping7x = OrderedDict({
+        mapping7x = {
             "my_index": {
                 "mappings": {
                     "properties": {
                         "city": {
                             "type": "text",
-                            "fields": {
-                                "keyword": {
-                                    "type": "keyword"
-                                }
-                            }
+                            "fields": {"keyword": {"type": "keyword"}},
                         }
                     }
                 }
             }
-        })
+        }
 
-        expected7x_source_only_false = {'city': ('text', None), 'city.keyword': ('keyword', None)}
-        expected7x_source_only_true = {'city': ('text', None)}
+        expected7x_source_only_false = {
+            "city": ("text", None),
+            "city.keyword": ("keyword", None),
+        }
+        expected7x_source_only_true = {"city": ("text", None)}
 
-        mapping6x = OrderedDict({
+        mapping6x = {
             "my_index": {
                 "mappings": {
                     "doc": {
                         "properties": {
                             "city": {
                                 "type": "text",
-                                "fields": {
-                                    "keyword": {
-                                        "type": "keyword"
-                                    }
-                                }
+                                "fields": {"keyword": {"type": "keyword"}},
                             }
                         }
                     }
                 }
             }
-        })
+        }
 
-        expected6x_source_only_false = {'city': ('text', None), 'city.keyword': ('keyword', None)}
-        expected6x_source_only_true = {'city': ('text', None)}
+        expected6x_source_only_false = {
+            "city": ("text", None),
+            "city.keyword": ("keyword", None),
+        }
+        expected6x_source_only_true = {"city": ("text", None)}
 
         # add a 5x mapping to get coverage of error
-        mapping5x = OrderedDict({
+        mapping5x = {
             "my_index": {
                 "mappings": {
                     "user": {
                         "properties": {
                             "name": {"type": "text"},
                             "user_name": {"type": "keyword"},
-                            "email": {"type": "keyword"}
+                            "email": {"type": "keyword"},
                         }
                     },
                     "tweet": {
                         "properties": {
                             "content": {"type": "text"},
                             "user_name": {"type": "keyword"},
-                            "tweeted_at": {"type": "date"}
+                            "tweeted_at": {"type": "date"},
                         }
-                    }
+                    },
                 }
             }
-        })
+        }
 
         result7x = FieldMappings._extract_fields_from_mapping(mapping7x)
         assert expected7x_source_only_false == result7x
 
-        result7x = FieldMappings._extract_fields_from_mapping(mapping7x, source_only=True)
+        result7x = FieldMappings._extract_fields_from_mapping(
+            mapping7x, source_only=True
+        )
         assert expected7x_source_only_true == result7x
 
         result6x = FieldMappings._extract_fields_from_mapping(mapping6x)
         assert expected6x_source_only_false == result6x
 
-        result6x = FieldMappings._extract_fields_from_mapping(mapping6x, source_only=True)
+        result6x = FieldMappings._extract_fields_from_mapping(
+            mapping6x, source_only=True
+        )
         assert expected6x_source_only_true == result6x
 
         with pytest.raises(NotImplementedError):
