@@ -22,11 +22,9 @@ from eland.tests.common import TestData
 
 
 class TestMetricSourceFields(TestData):
-
     def test_flights_all_metric_source_fields(self):
         ed_field_mappings = ed.FieldMappings(
-            client=ed.Client(ES_TEST_CLIENT),
-            index_pattern=FLIGHTS_INDEX_NAME
+            client=ed.Client(ES_TEST_CLIENT), index_pattern=FLIGHTS_INDEX_NAME
         )
         pd_flights = self.pd_flights()
 
@@ -36,40 +34,50 @@ class TestMetricSourceFields(TestData):
         assert pd_metric.dtypes.to_list() == ed_dtypes
         assert pd_metric.columns.to_list() == ed_fields
         assert len(es_date_formats) == len(ed_dtypes)
-        assert set(es_date_formats) == set([None])
+        assert set(es_date_formats) == {None}
 
     def test_flights_all_metric_source_fields_and_bool(self):
         ed_field_mappings = ed.FieldMappings(
-            client=ed.Client(ES_TEST_CLIENT),
-            index_pattern=FLIGHTS_INDEX_NAME
+            client=ed.Client(ES_TEST_CLIENT), index_pattern=FLIGHTS_INDEX_NAME
         )
         pd_flights = self.pd_flights()
 
-        ed_dtypes, ed_fields, es_date_formats = ed_field_mappings.metric_source_fields(include_bool=True)
-        pd_metric = pd_flights.select_dtypes(include=[np.number, 'bool'])
+        ed_dtypes, ed_fields, es_date_formats = ed_field_mappings.metric_source_fields(
+            include_bool=True
+        )
+        pd_metric = pd_flights.select_dtypes(include=[np.number, "bool"])
 
         assert pd_metric.dtypes.to_list() == ed_dtypes
         assert pd_metric.columns.to_list() == ed_fields
         assert len(es_date_formats) == len(ed_dtypes)
-        assert set(es_date_formats) == set([None])
+        assert set(es_date_formats) == {None}
 
     def test_flights_all_metric_source_fields_bool_and_timestamp(self):
         ed_field_mappings = ed.FieldMappings(
-            client=ed.Client(ES_TEST_CLIENT),
-            index_pattern=FLIGHTS_INDEX_NAME
+            client=ed.Client(ES_TEST_CLIENT), index_pattern=FLIGHTS_INDEX_NAME
         )
         pd_flights = self.pd_flights()
 
-        ed_dtypes, ed_fields, es_date_formats = ed_field_mappings.metric_source_fields(include_bool=True, include_timestamp=True)
-        pd_metric = pd_flights.select_dtypes(include=[np.number, 'bool', 'datetime'])
+        ed_dtypes, ed_fields, es_date_formats = ed_field_mappings.metric_source_fields(
+            include_bool=True, include_timestamp=True
+        )
+        pd_metric = pd_flights.select_dtypes(include=[np.number, "bool", "datetime"])
 
         assert pd_metric.dtypes.to_list() == ed_dtypes
         assert pd_metric.columns.to_list() == ed_fields
         assert len(es_date_formats) == len(ed_dtypes)
-        assert set(es_date_formats) == set({"strict_date_hour_minute_second", None}) # TODO - test position of date_format
+        assert set(es_date_formats) == set(
+            {"strict_date_hour_minute_second", None}
+        )  # TODO - test position of date_format
 
     def test_ecommerce_selected_non_metric_source_fields(self):
-        field_names = ['category', 'currency', 'customer_birth_date', 'customer_first_name', 'user']
+        field_names = [
+            "category",
+            "currency",
+            "customer_birth_date",
+            "customer_first_name",
+            "user",
+        ]
         """
         Note: non of there are metric
         category                       object
@@ -81,7 +89,7 @@ class TestMetricSourceFields(TestData):
         ed_field_mappings = ed.FieldMappings(
             client=ed.Client(ES_TEST_CLIENT),
             index_pattern=ECOMMERCE_INDEX_NAME,
-            display_names=field_names
+            display_names=field_names,
         )
         pd_ecommerce = self.pd_ecommerce()[field_names]
 
@@ -94,7 +102,14 @@ class TestMetricSourceFields(TestData):
         assert set(es_date_formats) == set()
 
     def test_ecommerce_selected_mixed_metric_source_fields(self):
-        field_names = ['category', 'currency', 'customer_birth_date', 'customer_first_name', 'total_quantity', 'user']
+        field_names = [
+            "category",
+            "currency",
+            "customer_birth_date",
+            "customer_first_name",
+            "total_quantity",
+            "user",
+        ]
         """
         Note: one is metric
         category                       object
@@ -107,20 +122,20 @@ class TestMetricSourceFields(TestData):
         ed_field_mappings = ed.FieldMappings(
             client=ed.Client(ES_TEST_CLIENT),
             index_pattern=ECOMMERCE_INDEX_NAME,
-            display_names=field_names
+            display_names=field_names,
         )
         pd_ecommerce = self.pd_ecommerce()[field_names]
 
         ed_dtypes, ed_fields, es_date_formats = ed_field_mappings.metric_source_fields()
         pd_metric = pd_ecommerce.select_dtypes(include=np.number)
         assert len(es_date_formats) == len(ed_dtypes)
-        assert set(es_date_formats) == set([None])
+        assert set(es_date_formats) == {None}
 
         assert pd_metric.dtypes.to_list() == ed_dtypes
         assert pd_metric.columns.to_list() == ed_fields
 
     def test_ecommerce_selected_all_metric_source_fields(self):
-        field_names = ['total_quantity', 'taxful_total_price', 'taxless_total_price']
+        field_names = ["total_quantity", "taxful_total_price", "taxless_total_price"]
         """
         Note: all are metric
         total_quantity           int64
@@ -130,7 +145,7 @@ class TestMetricSourceFields(TestData):
         ed_field_mappings = ed.FieldMappings(
             client=ed.Client(ES_TEST_CLIENT),
             index_pattern=ECOMMERCE_INDEX_NAME,
-            display_names=field_names
+            display_names=field_names,
         )
         pd_ecommerce = self.pd_ecommerce()[field_names]
 
@@ -140,4 +155,4 @@ class TestMetricSourceFields(TestData):
         assert pd_metric.dtypes.to_list() == ed_dtypes
         assert pd_metric.columns.to_list() == ed_fields
         assert len(es_date_formats) == len(ed_dtypes)
-        assert set(es_date_formats) == set([None])
+        assert set(es_date_formats) == {None}
