@@ -21,10 +21,9 @@ from eland.tests.common import assert_pandas_eland_series_equal
 
 
 class TestSeriesHeadTail(TestData):
-
     def test_head_tail(self):
-        pd_s = self.pd_flights()['Carrier']
-        ed_s = ed.Series(ES_TEST_CLIENT, FLIGHTS_INDEX_NAME, 'Carrier')
+        pd_s = self.pd_flights()["Carrier"]
+        ed_s = ed.Series(ES_TEST_CLIENT, FLIGHTS_INDEX_NAME, "Carrier")
 
         pd_s_head = pd_s.head(10)
         ed_s_head = ed_s.head(10)
