@@ -23,7 +23,6 @@ from eland.tests.common import TestData
 
 
 class TestSeriesFrameHist(TestData):
-
     def test_flight_delay_min_hist(self):
         pd_flights = self.pd_flights()
         ed_flights = self.ed_flights()
@@ -31,14 +30,12 @@ class TestSeriesFrameHist(TestData):
         num_bins = 10
 
         # pandas data
-        pd_flightdelaymin = np.histogram(pd_flights['FlightDelayMin'], num_bins)
+        pd_flightdelaymin = np.histogram(pd_flights["FlightDelayMin"], num_bins)
 
-        pd_bins = pd.DataFrame(
-            {'FlightDelayMin': pd_flightdelaymin[1]})
-        pd_weights = pd.DataFrame(
-            {'FlightDelayMin': pd_flightdelaymin[0]})
+        pd_bins = pd.DataFrame({"FlightDelayMin": pd_flightdelaymin[1]})
+        pd_weights = pd.DataFrame({"FlightDelayMin": pd_flightdelaymin[0]})
 
-        ed_bins, ed_weights = ed_flights['FlightDelayMin']._hist(num_bins=num_bins)
+        ed_bins, ed_weights = ed_flights["FlightDelayMin"]._hist(num_bins=num_bins)
 
         # Numbers are slightly different
         print(pd_bins, ed_bins)
@@ -52,17 +49,19 @@ class TestSeriesFrameHist(TestData):
         num_bins = 10
 
         # pandas data
-        pd_filteredhist = np.histogram(pd_flights[pd_flights.FlightDelay == True].FlightDelayMin, num_bins)
+        pd_filteredhist = np.histogram(
+            pd_flights[pd_flights.FlightDelay == True].FlightDelayMin, num_bins
+        )
 
-        pd_bins = pd.DataFrame(
-            {'FlightDelayMin': pd_filteredhist[1]})
-        pd_weights = pd.DataFrame(
-            {'FlightDelayMin': pd_filteredhist[0]})
+        pd_bins = pd.DataFrame({"FlightDelayMin": pd_filteredhist[1]})
+        pd_weights = pd.DataFrame({"FlightDelayMin": pd_filteredhist[0]})
 
         d = ed_flights[ed_flights.FlightDelay == True].FlightDelayMin
         print(d.info_es())
 
-        ed_bins, ed_weights = ed_flights[ed_flights.FlightDelay == True].FlightDelayMin._hist(num_bins=num_bins)
+        ed_bins, ed_weights = ed_flights[
+            ed_flights.FlightDelay == True
+        ].FlightDelayMin._hist(num_bins=num_bins)
 
         # Numbers are slightly different
         assert_almost_equal(pd_bins, ed_bins)
@@ -70,4 +69,4 @@ class TestSeriesFrameHist(TestData):
 
     def test_invalid_hist(self):
         with pytest.raises(ValueError):
-            assert self.ed_ecommerce()['products.tax_amount'].hist()
+            assert self.ed_ecommerce()["products.tax_amount"].hist()
