@@ -20,10 +20,9 @@ from eland.tests.common import TestData
 
 
 class TestSeriesRepr(TestData):
-
     def test_repr_flights_carrier(self):
-        pd_s = self.pd_flights()['Carrier']
-        ed_s = ed.Series(ES_TEST_CLIENT, FLIGHTS_INDEX_NAME, 'Carrier')
+        pd_s = self.pd_flights()["Carrier"]
+        ed_s = ed.Series(ES_TEST_CLIENT, FLIGHTS_INDEX_NAME, "Carrier")
 
         pd_repr = repr(pd_s)
         ed_repr = repr(ed_s)
@@ -31,8 +30,8 @@ class TestSeriesRepr(TestData):
         assert pd_repr == ed_repr
 
     def test_repr_flights_carrier_5(self):
-        pd_s = self.pd_flights()['Carrier'].head(5)
-        ed_s = ed.Series(ES_TEST_CLIENT, FLIGHTS_INDEX_NAME, 'Carrier').head(5)
+        pd_s = self.pd_flights()["Carrier"].head(5)
+        ed_s = ed.Series(ES_TEST_CLIENT, FLIGHTS_INDEX_NAME, "Carrier").head(5)
 
         pd_repr = repr(pd_s)
         ed_repr = repr(ed_s)
